+++ app/src/test/java/com/github/aedge90/nmm/StrategyRunnableTest.java
@@ -67,7 +67,7 @@ public class StrategyRunnableTest {
         assertEquals(0, possibleMoves.size());
 
         //minimizing players worst case is MAX. black is the maximizing player
-        int result = strategy.evaluation(mPlayerWhite, possibleMoves,0);
+        double result = strategy.evaluation(mPlayerWhite, possibleMoves,0);
         assertEquals(strategy.MAX, result);
     }
